refactor: extract shared checkpoint and model utilities
Move algorithm-agnostic checkpoint functions (download_checkpoint_from_hub,
ensure_checkpoint_is_local, load_checkpoint_config, load_checkpoint_weights)
from convert.eagle.utils to a new convert.utils module. Move
conditional_torch_compile from models.eagle3.core to models.utils.

Update all consumers (eagle converters, MTP converter, eagle3/mtp/peagle
cores) to import from the shared modules. Update test patch targets in
test_eagle_utils.py and test_eagle3_converter.py accordingly.
